+++ bot/__init__.py
@@ -8,4 +8,4 @@ You should have received a copy of the license along with this
 work. If not, see <http://creativecommons.org/licenses/by/4.0/>.
 """
 
-ALL = ["stickfix", "database", "utils"]
+__all__ = ["stickfix", "database", "utils"]