+++ db/__init__.py
@@ -1,0 +1,7 @@
+""" Stickfix's database module initializer
+"""
+
+__author__ = "Ignacio Slater Muñoz <ignacio.slater@ug.uchile.cl>"
+__version__ = "3.0.0000"
+
+ALL = ["sf_database"]